agents-quickstart-pocs/strands-sdk/ml-feature-analyzer-agent-using-strands/requirements.txt
# Core AWS and ML dependencies
boto3>=1.34.0
sagemaker>=2.237.3
sagemaker-core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.5.0

# GenAI and Agents
anthropic>=0.25.0
strands-agents>=0.1.0

# Infrastructure
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# CLI and Interface
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0
streamlit>=1.37.0
python-dotenv>=1.0.0

# Development and Testing
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.3.0
flake8>=6.0.0
mypy>=1.8.0

# Jupyter (for analysis notebooks)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
